pyserver/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pillow==10.2.0
hidapi==0.14.0
pyserial==3.5
psutil==5.9.8
requests==2.31.0
aiohttp==3.9.3
python-multipart==0.0.9
pydantic>=2.0.0
svglib==1.5.1
reportlab==4.1.0
